# shared/api/order/api.yaml
openapi: 3.0.0
info:
  title: OrderService API
  version: 1.0.0
paths:
  /api/v1/orders:
    post:
      description: >-
        Создание нового заказа на основе выбранных пользователем деталей.
      tags:
        - Order
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                user_uuid:
                  $ref: "#/components/schemas/UserUuid"
                part_uuids:
                  $ref: "#/components/schemas/PartUuids"
      responses:
        '200':
          description: Успешно создан заказ
          content:
            application/json:
              schema:
                type: object
                properties:
                  order_uuid:
                    $ref: "#/components/schemas/OrderUuid"
                  total_price:
                    $ref: "#/components/schemas/TotalPrice"
        '400':
          $ref: "#/components/responses/400"
        '401':
          $ref: "#/components/responses/401"
        '500':
          $ref: "#/components/responses/5xx"
  /api/v1/orders/{order_uuid}/pay:
    post:
      description: >-
        Проведение оплаты ранее созданного заказ
      tags:
        - Order
      parameters:
        - name: order_uuid
          schema:
            $ref: "#/components/schemas/OrderUuid"
          in: path
          required: true
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                payment_method:
                  $ref: "#/components/schemas/PaymentMethod"
      responses:
        '200':
          description: Успешно выполнена оплата заказа
          content:
            application/json:
              schema:
                type: object
                properties:
                  transaction_uuid:
                    $ref: "#/components/schemas/TransactionUuid"
        '400':
          $ref: "#/components/responses/400"
        '401':
          $ref: "#/components/responses/401"
        '404':
          description: не найден заказ
        '500':
          $ref: "#/components/responses/5xx"
  /api/v1/orders/{order_uuid}:
    get:
      description: >-
        Возвращает информацию о заказе.
      tags:
        - Order
      parameters:
        - name: order_uuid
          schema:
            $ref: "#/components/schemas/OrderUuid"
          required: true
          in: path
      responses:
        '200':
          description: Успешно выполнен запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        '400':
          $ref: "#/components/responses/400"
        '401':
          $ref: "#/components/responses/401"
        '404':
          $ref: "#/components/responses/404"
        '500':
          $ref: "#/components/responses/5xx"
  /api/v1/orders/{order_uuid}/cancel:
    post:
      parameters:
        - name: order_uuid
          schema:
            $ref: "#/components/schemas/OrderUuid"
          required: true
          in: path
      responses:
        '204':
          description: заказ успешно отменён
        '400':
            $ref: "#/components/responses/400"
        '404':
          $ref: "#/components/responses/404"
        '409':
          description: заказ уже оплачен и не может быть отменён
        '500':
          $ref: "#/components/responses/5xx"
components:
  responses:
    '404':
      description: не найден заказ
    '400':
      description: Невалидные данные ввода
    '401':
      description: Неавторизованный доступ
    5xx:
      description: Ошибка сервера
      content:
        application/json:
          schema:
            type: object
            required:
              - message
            properties:
              message:
                type: string
                description: Описание ошибки
                example: что-то пошло не так
              request_id:
                type: string
                description: >-
                  Идентификатор запроса. Предназначен для более быстрого поиска проблем.
                example: g12ugs67gqw67yu12fgeuqwd
              code:
                type: integer
                description: >-
                  Код ошибки. Предназначен для классификации проблем и более быстрого решения проблем.
                example: 12345
  schemas:
    OrderUuid:
      type: string
      format: uuid
      description: Уникальный идентификатор заказа
      example: 'cae36e0f-69e5-4fa8-a179-a52d083c5549'
    UserUuid:
      type: string
      format: uuid
      description: Уникальный идентификатор пользователя
      example: 'cae36e0f-69e5-4fa8-a179-a52d083c5549'
    PartUuids:
      type: array
      description: Список UUID деталей
      items:
        type: string
      example: [ "string", "string" ]
    TotalPrice:
      type: number
      format: double
      description: Итоговая стоимость
      example: 1900.29
    TransactionUuid:
      type: string
      format: uuid
      description: UUID транзакции (если оплачен)
      example: "string"
    PaymentMethod:
      type: string
      enum: [ UNKNOWN, CARD, SBP, CREDIT_CARD, INVESTOR_MONEY ]
      description: Способ оплаты (если оплачен)
      example: "CARD"
    Status:
      type: string
      enum: [ PENDING_PAYMENT, PAID, CANCELLED ]
      description: Статус (PENDING_PAYMENT, PAID, CANCELLED)
      example: "PAID"
    Order:
      type: object
      description: Заказ
      required:
        - id
      properties:
        order_uuid:
          $ref: "#/components/schemas/OrderUuid"
        user_uuid:
          $ref: "#/components/schemas/UserUuid"
        part_uuids:
          $ref: "#/components/schemas/PartUuids"
        transaction_uuid:
          $ref: "#/components/schemas/TransactionUuid"
        payment_method:
          $ref: "#/components/schemas/PaymentMethod"
        status:
          $ref: "#/components/schemas/Status"
        total_price:
          $ref: "#/components/schemas/TotalPrice"
tags:
  - name: Order
    description: Сервис заказов управляет заказами на постройку космических кораблей